Remove sha2-soc files
